Fix unexpected cast int to uint

+++ com.unity.visualeffectgraph/Editor/Expressions/VFXExpressionMesh.cs
@@ -167,7 +167,7 @@ namespace UnityEditor.VFX
         {
             var meshReduce = constParents[0];
             var mesh = meshReduce.Get<Mesh>();
-            return VFXValue.Constant(mesh.vertexCount);
+            return VFXValue.Constant((uint)mesh.vertexCount);
         }
     }
